# DepOpenSSL, initialize