--- a/worker/src/DepOpenSSL.cpp
+++ b/worker/src/DepOpenSSL.cpp
@@ -22,6 +22,7 @@ void DepOpenSSL::ClassInit()
 	  []
 	  {
 		  MS_DEBUG_TAG(info, "openssl version: \"%s\"", OpenSSL_version(OPENSSL_VERSION));
+		  MS_DEBUG_TAG(info, "openssl CPU info: \"%s\"", OpenSSL_version(OPENSSL_CPU_INFO));
 
 		  // Initialize some crypto stuff.
 		  RAND_poll();