@ibc 2023-11-16
Do we need these before/after logs permanently? I assume we don't once we know how it works.