@jmillan 2023-11-16
So it seems the caps are loaded without the need to call `OPENSSL_init_ssl()`.

Still we are not 100% sure HW encryption is being used.

I'm keeping this PR open for a while just for discussion.