@jmillan 2023-11-16
Please, see the `Edit` section of the issue description. We have an easy way to know if HW encryption is being used.

I've kept just a single log line in this PR, which I'll commit. I've removed the OpenSSL initialization as it seems it's not needed.